+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.2.0 - master
 - Added `cornerCircle` to create a full circle based on the height of the view
 - Added more examples and documentation
-- Added `DezignableBackground`
+- Added `DezignableBackground` with background fill color and gradient background
 
 ## 1.1.0 - 26 feb 2016
 - Mostly refactored to protocol based extensions